refactor(config): extract config-section builders → thin resolveSmartServerConfig (<500)
Move the 6 cohesive field-builders out of resolveSmartServerConfig's object
literal into resolve-config-sections.ts (resolveLlmSection, resolveRagSection,
resolveMcpSection, resolveAgentSection, resolvePromptsSection,
resolvePipelineSelection). Promote the duplicated `get` accessor to
yaml-loader.ts as a shared export; remove both local copies. config.ts drops
from 526 → 269 lines; resolve-config-sections.ts is 303 lines. Behavior-
preserving: public API and all 9 config test suites unchanged (134/134 pass).
